refactor: move InfoBanner to new package
